+++ packages/ui/src/form/primitives/searchable-chip-select/test/index.test.tsx
@@ -176,7 +176,7 @@ describe( 'SearchableChipSelect', () => {
 			expect.anything()
 		);
 		expect(
-			screen.getByRole( 'button', { name: 'Remove' } )
+			screen.getByRole( 'button', { name: 'Remove Strawberry' } )
 		).toBeVisible();
 	} );